+++ changelog.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - internal documentation remastering
   added usage examples, components specification.
 
+### Fixed
+
+- `Format`, `KeepStream` components instance representation fixed
+
 ## [3.1.0] - 2023-01-25
 ### Changed